fin implementation fontionalitées instantSearch + reorganisation fichiers
