+++ client/src/pages/Categories.jsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { Switch, Route, useRouteMatch, useParams } from 'react-router-dom'
 import algoliasearch from 'algoliasearch/lite'
-import { InstantSearch, Configure, SearchBox, InfiniteHits } from 'react-instantsearch-dom'
+import { InstantSearch, Configure, SearchBox, InfiniteHits, SortBy } from 'react-instantsearch-dom'
+import Filters from '../components/Filters'
 
 const Category = () => {
   const searchClient = algoliasearch('H2YGA5NBNG', '26b1badfb8cb68a219e300f7cd17df1e')
@@ -15,8 +16,17 @@ const Category = () => {
       )}
 
       <InstantSearch searchClient={searchClient} indexName='demo_ecommerce'>
-        <Configure filters={'brand:' + categoryName} />
+        <Configure filters={'categories:"' + categoryName + '"'} />
         <SearchBox />
+        <Filters />
+        <SortBy
+          items={[
+            { value: 'demo_ecommerce', label: 'Price desc' },
+            { value: 'price_asc', label: 'Price asc' },
+            { value: 'popularity', label: 'Popularity' }
+          ]}
+          defaultRefinement='popularity'
+        />
         <InfiniteHits />
       </InstantSearch>
